DDoS_12/requirements.txt
annotated-types==0.7.0
appdirs==1.4.4
asgiref==3.8.1
attrs==24.3.0
Automat==24.8.1
blinker==1.9.0
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.2
cryptography==44.0.0
cssselect==1.2.0
cycler==0.12.1
defusedxml==0.7.1
distlib==0.3.9
Django==5.1.4
django-allauth==65.9.0
djangorestframework==3.16.0
filelock==3.16.1
Flask==3.1.2
fonttools==4.57.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.0
kiwisolver==1.4.8
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.2.5
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
parsel==1.9.1
pillow==11.1.0
platformdirs==4.3.6
Protego==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyDispatcher==2.0.7
PyJWT==2.10.1
pyOpenSSL==24.3.0
pyparsing==3.2.3
pyshark==0.6
python-dateutil==2.9.0.post0
pytz==2025.2
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
rsa==4.9.1
scapy==2.6.1
scikit-learn==1.6.1
scipy==1.15.2
Scrapy==2.12.0
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
termcolor==3.0.1
threadpoolctl==3.6.0
tldextract==5.1.3
tqdm==4.67.1
Twisted==24.11.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.29.1
w3lib==2.2.1
waitress==3.0.2
Werkzeug==3.1.3
xgboost==3.0.0
zope.interface==7.2
